+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-1.0.7.0 (unreleased)
+1.1.0.0 (unreleased)
 --------------------
 - Fix handling of product applicable to backorders
 - Fix handling of divisible products